cleaner implementation for input_add
